Move bson files to core utils
