+++ src/realm/object-store/CMakeLists.txt
@@ -110,14 +110,7 @@ if(REALM_ENABLE_SYNC)
         sync/impl/sync_client.hpp
         sync/impl/sync_file.hpp
         sync/impl/sync_metadata.hpp
-        sync/impl/network_reachability.hpp
-
-        util/bson/bson.hpp
-        util/bson/min_key.hpp
-        util/bson/max_key.hpp
-        util/bson/regular_expression.hpp
-        util/bson/indexed_map.hpp
-        util/bson/mongo_timestamp.hpp)
+        sync/impl/network_reachability.hpp)
 
     list(APPEND SOURCES
         sync/app.cpp
@@ -133,9 +126,7 @@ if(REALM_ENABLE_SYNC)
         sync/mongo_database.cpp
         sync/push_client.cpp
         sync/impl/sync_file.cpp
-        sync/impl/sync_metadata.cpp
-        util/bson/bson.cpp
-        util/bson/regular_expression.cpp)
+        sync/impl/sync_metadata.cpp)
     if(APPLE)
         list(APPEND HEADERS
             sync/impl/apple/network_reachability_observer.hpp
