+++ src/realm/object-store/sync/app_credentials.cpp
@@ -17,7 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 #include <realm/object-store/sync/app_credentials.hpp>
-#include <realm/object-store/util/bson/bson.hpp>
+#include <realm/util/bson/bson.hpp>
 
 namespace realm::app {
