+++ test/object-store/sync/flx_sync.cpp
@@ -37,7 +37,7 @@
 #include <realm/object-store/sync/mongo_database.hpp>
 #include <realm/object-store/sync/mongo_collection.hpp>
 #include <realm/object-store/sync/async_open_task.hpp>
-#include <realm/object-store/util/bson/bson.hpp>
+#include <realm/util/bson/bson.hpp>
 #include <realm/object-store/sync/sync_session.hpp>
 
 #include <realm/sync/client_base.hpp>
